Fix attributes being lost when an existing device is re-added
INSERT OR REPLACE deletes and inserts when the record already exists.
Since table constraints have ON DELETE CASCADE, the attributes (and probably
endpoints and so on) were being deleted when an existing device is saved.
After rebooting HA, no quirk would match since the device would have not
manufacturer attr.